counter: value_type has been changed to int64_t.

+++ include/handystats/metrics/counter.hpp
@@ -14,7 +14,7 @@ class counter
 {
 public:
 
-	typedef long long value_type;
+	typedef int64_t value_type;
 	typedef typename chrono::default_duration time_duration;
 	typedef typename chrono::steady_clock<time_duration> clock;
 	typedef typename clock::time_point time_point;
